So many performance tweaks, call me Viagra
Fixed lagging issues with pretty much everything. TLDR, collision
generated for the sprites was like fucked up and had like 500 triangles
for it's collision when most items needed only like 5-10. I just reset
the collision and it regenerated another collision with that many less
triangles. Secondly, tweaked the code to prevent a little searching for
gameObjects by name as possible, prevented a shit ton of "new" classes
from being generated (for example when you fire a bullet, no need to
generate another weapon script when you don't need to). Also all the
sprites were rendering at like 2048p resolution, so for many of the
sprites such as Food where I lowered that to like 256p and it looks
exactly the same.